rename trace wrapper and do lazy decoding
