+++ src/mdio/segy/_workers.py
@@ -12,7 +12,7 @@ from segy import SegyFile
 
 from mdio.api.io import to_mdio
 from mdio.builder.schemas.dtype import ScalarType
-from mdio.segy._disaster_recovery_wrapper import SegyFileTraceDataWrapper
+from mdio.segy._raw_trace_wrapper import SegyFileRawTraceWrapper
 
 if TYPE_CHECKING:
     from segy.arrays import HeaderArray
@@ -139,7 +139,7 @@ def trace_worker(  # noqa: PLR0913
     # For that reason, we have wrapped the accessors to provide an interface that can be removed
     # and not require additional changes to the below code.
     # NOTE: The `raw_header_key` code block should be removed in full as it will become dead code.
-    traces = SegyFileTraceDataWrapper(segy_file, live_trace_indexes)
+    traces = SegyFileRawTraceWrapper(segy_file, live_trace_indexes)
 
     ds_to_write = dataset[worker_variables]
